New HRI framework/new_framework_microservices/question_service/requirements.txt
Flask
tensorflow==1.14.0
Keras==2.3.1
statistics
pandas
matplotlib
pillow
python-logstash
gunicorn==19.10.0
py_eureka_client==0.7.4
python-json-logger
llvmlite==0.31.0
librosa==0.7.1
